refactor: move macro to its own file
